fix: V-001 security vulnerability
Automated security fix generated by Orbis Security AI

+++ src/ctest.c
@@ -620,7 +620,7 @@ static char* ctest_vsprintf_char(const char* format, va_list va)
     }
     else
     {
-        result = malloc(neededSize + 1);
+        result = malloc((size_t)neededSize + 1);
         if (result == NULL)
         {
             LogError("failure in malloc");
@@ -628,7 +628,7 @@ static char* ctest_vsprintf_char(const char* format, va_list va)
         }
         else
         {
-            if (vsnprintf(result, neededSize + 1, format, va) != neededSize)
+            if (vsnprintf(result, (size_t)neededSize + 1, format, va) != neededSize)
             {
                 LogError("inconsistent vsnprintf behavior format, neededSize=%d + 1, format=%s, va=%p", neededSize, format, (void*)&va);
                 free(result);
